requirements.txt
google-cloud-pubsub==2.15.0
google-cloud-datastore==2.15.2